backend/client-upload/requirements.txt
psycopg2-binary>=2.9.11
boto3>=1.42.25
